+++ libraries/bio/src/molecule-3d/molecule-3d-units-handler.ts
@@ -5,6 +5,7 @@ import * as ui from 'datagrok-api/ui';
 import {BuildDataFunc, UnitsHandlerBase} from '../utils/units-handler-base';
 import {Molecule3DBase, Molecule3DBuildDataFunc} from './types';
 import {buildDataMmcif, buildDataPdb, buildDataPdbqt} from './molecule-3d-build-data';
+import {IPdbHelper} from '../pdb/pdb-helper';
 
 export const Temps = new class {
   uh = `units-handler.${DG.SEMTYPE.MOLECULE3D}`;
@@ -43,6 +44,30 @@ export class Molecule3DUnitsHandler extends UnitsHandlerBase<any, Molecule3DBase
 
   protected getBuildDataFunc(): BuildDataFunc<any, Molecule3DBase> {
     return Molecule3DBuildDataFuncs[this.units];
+  }
+
+  public async getAsPdb(ph: IPdbHelper): Promise<DG.Column<string>> {
+    if (this.units === 'pdb') return this.column;
+    let convertToPdb: (src: string) => Promise<string>;
+    switch (this.units) {
+      case 'pdbqt': {
+        convertToPdb = (srcPdbqt: string): Promise<string> => {
+          return ph.pdbqtToMol(srcPdbqt);
+        };
+        break;
+      }
+      default:
+        throw new Error('Unsupported units \'${this.units}\' of the Molecule3D column');
+    }
+
+    const colLength: number = this.column.length;
+    const resCol: DG.Column = DG.Column.fromType(DG.TYPE.STRING, this.column.name, colLength);
+    for (let rowI = 0; rowI < colLength; ++rowI) {
+      const srcVal = this.column.get(rowI);
+      const pdbVal = srcVal === null ? null : await convertToPdb(srcVal);
+      resCol.set(rowI, pdbVal);
+    }
+    return resCol;
   }
 
   public static getOrCreate(col: DG.Column): Molecule3DUnitsHandler {